% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_meta}
\alias{check_meta}
\title{Check that the metadata provided is what we expect}
\usage{
check_meta(
  meta,
  check_replicates = FALSE,
  group_variable = "Group",
  time_variable = "Timepoint"
)
}
\arguments{
\item{meta}{metadata}

\item{check_replicates}{boolean, optional, default: FALSE
If TRUE, checks whether metadata contains a column Replicate}
}
\value{
meta returns the metadata with additional columns if necessary.
}
\description{
This method will raise errors if the metadata provided is not as expected.
}
\keyword{internal}
