% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_validation.R
\name{consensus_matrix}
\alias{consensus_matrix}
\title{Compute consensus matrix from labels}
\usage{
consensus_matrix(labels, scale = TRUE)
}
\arguments{
\item{labels}{matrix (n_genes, n_clusters)
Matrix containing the set of labels}

\item{scale}{boolean, optional, default: TRUE
Whether to rescale the matrix}
}
\description{
Compute consensus matrix from labels
}
