% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_terms_analysis.R
\name{create_go_term_mapping}
\alias{create_go_term_mapping}
\title{Create the Gene to GO Term mapping}
\usage{
create_go_term_mapping(genes, gene_col = "refseq_mrna")
}
\arguments{
\item{genes}{dataframe, with `gene_col` and a column corresponding to the
`go_id`}

\item{gene_col}{the column of the genes data frame that contains the
correct gene reference. By default, is refseq_mrna}
}
\description{
Create the Gene to GO Term mapping
}
