% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moanin_model.R
\name{create_moanin_model}
\alias{create_moanin_model}
\title{Create a moanin model}
\usage{
create_moanin_model(meta, formula = NULL, basis = NULL,
  degrees_of_freedom = 4)
}
\arguments{
\item{meta}{data.frame containing the metadata. Metadata needs to contain
column "Group" and "Timepoint"}

\item{formula}{formula object, optional, default: NUlL
Formula of the splines, e.g:
    formula = ~Group:splines::ns(Timepoint, df=4) + Group + 0}

\item{basis}{matrix, optional, default: NULL
A basis object.}

\item{degrees_of_freedom}{int, optional, default: 4
Number of degrees of freedom to use if neither the basis nor the
formula is provided}
}
\value{
An object containing the following information:

- The basis: a matrix of shape (degrees_of_freedom, n_samples)
- The metadata: a data.frame of shape (n_samples, n_metadat)
- The formula used, when provided.
- The number of degrees of freedom.
}
\description{
This model is used to conserve information relating to the model, such as
the formula, the basis, the meta data.
}
\examples{
# Load some data
data(shoemaker2015)
meta = shoemaker2015$meta

# Use the default options
moanin = create_moanin_model(meta)
print(dim(moanin$basis))

# Change the number of degrees of freedom
moanin = create_moanin_model(meta, degrees_of_freedom=6)
print(dim(moanin$basis))
}
