% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_timepoints.R
\name{create_timepoints_contrasts}
\alias{create_timepoints_contrasts}
\title{Creates pairwise contrasts for all timepoints}
\usage{
create_timepoints_contrasts(group1, group2, moanin_model)
}
\arguments{
\item{group1}{First group to consider}

\item{group2}{Second group to consider}

\item{moanin_model}{object of class \code{moanin_model}, an object containing
all related information to the splines model used. See
\code{\link{create_moanin_model}} for more details.}
}
\description{
Creates pairwise contrasts for all timepoints
}
\examples{
data(shoemaker2015)
meta = shoemaker2015$meta
moanin_model = moanin::create_moanin_model(meta)
contrasts = create_timepoints_contrasts("C", "M", moanin_model)

}
