% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_timepoints.R
\name{DE_timepoints}
\alias{DE_timepoints}
\title{Fit weekly differential expression analysis}
\usage{
DE_timepoints(data, splines_model, contrasts, use_voom_weights = TRUE)
}
\arguments{
\item{data}{Gene expression data}

\item{splines_model}{splines_model}

\item{contrasts}{Contrast to use.}

\item{use_voom_weights}{boolean: whether to use voom weights or not}
}
\description{
Fit weekly differential expression analysis
}
