% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Small data set for running examples}
\format{
Two objects are loaded, a data frame of expression of 500 genes by 84
  samples (\code{testData}), and a data frame with meta information on those
  84 samples (\code{testMeta}).
}
\description{
Small data set for running examples
}
\details{
This data is a subset of the full time course data available as
  \code{\link[timecoursedata]{shoemaker2015}} and is only provided for the
  purpose of running examples, and not for biological meaning. Users should
  refer to the full data set.
}
\examples{
#code used to create data:
\dontrun{
library(timecoursedata)
data(shoemaker2015)
testData<-shoemaker2015$data[1:500,]
whSamples<-which(shoemaker2015$meta$Group \%in\% c("C","K"))
testData<-testData[,whSamples]
testMeta<-shoemaker2015$meta[whSamples,]
save(list=c("testData","testMeta"),file="data/exampleData.rda")
}
}
\keyword{data}
