% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{expression_filtering}
\alias{expression_filtering}
\title{Utility function to filter out low-expressed genes}
\usage{
expression_filtering(counts, min_counts = 20, min_samples = 3)
}
\arguments{
\item{counts}{n by p matrix containing the count data.}

\item{min_counts}{integer, corresponding to the minimum number of counts for
the gene to be considered expressed in a sample. Default is 20.}

\item{min_samples}{integer, corresponding to the minimum of samples for a
gene to be expressed to be included in downstream analyses.}
}
\value{
The filtered counts matrix
}
\description{
Utility function to filter out low-expressed genes
}
\keyword{internal}
