% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{fit_predict_splines}
\alias{fit_predict_splines}
\title{Fit and predict splines}
\usage{
fit_predict_splines(data, splines_model, weights = NULL,
  meta_prediction = NULL)
}
\arguments{
\item{data}{the data}

\item{splines_model}{splines_model}

\item{weights}{weights}
}
\value{
y_fitted the fitted y values
}
\description{
Fit and predict splines
}
