% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{fit_splines}
\alias{fit_splines}
\title{Fit splines}
\usage{
fit_splines(data, moanin_model, weights = NULL)
}
\arguments{
\item{data}{the data}

\item{moanin_model}{moanin_model}

\item{weights}{weigts}
}
\value{
beta coefficients
}
\description{
Fit splines
}
