% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_validation.R
\name{get_auc_similarity_scores}
\alias{get_auc_similarity_scores}
\title{Plot AUC consensus matrix}
\usage{
get_auc_similarity_scores(labels, method = "consensus")
}
\arguments{
\item{labels}{list of lables}
}
\description{
Plot AUC consensus matrix
}
