% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_timecourse.R, R/validation.R
\name{expand_contrast}
\alias{expand_contrast}
\alias{internal}
\alias{check_data_meta}
\alias{check_is_2d}
\alias{is_contrasts}
\title{Internal Validation Checks}
\usage{
expand_contrast(moanin_model, contrast_vector)

check_data_meta(data, object)

check_is_2d(X)

is_contrasts(contrasts, moanin_model)
}
\value{
Does not return anything. Only hits errors if there are problems.

\code{is_contrasts} returns the contrasts, with any corrections.
}
\description{
Will check that the contrasts provided are indeed contrasts. contrasts are
 expected to be either a vector of string or a matrix containing the
 contrasts coefficients.
}
\details{
If a vector of string is provided, the function will call
 limma::makeContrast in order to obtain the contrasts coefficients.

If a contrasts matrix is provided, it will perform a number of
  checks on the contrasts matrix to make sure it contains the number of rows
  expected, and that each contrast indeed sums to 0.
}
\keyword{internal}
