% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_validation.R
\name{plot_cdf_consensus}
\alias{plot_cdf_consensus}
\title{Plot CDF consensus matrix}
\usage{
plot_cdf_consensus(labels)
}
\arguments{
\item{labels}{list of labels}
}
\description{
Plot CDF consensus matrix
}
