% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_centroids}
\alias{plot_centroids}
\title{Plotting centroids}
\usage{
plot_centroids(centroids, splines_model, colors = NULL, smooth = FALSE)
}
\arguments{
\item{centroids}{matrix (k, t) containing the centroids}

\item{splines_model}{splines_model}

\item{colors}{vector, optional, default NULL
vector of colors}

\item{smooth}{boolean, optional, default: FALSE
Whether to smooth the centroids or not.}

\item{meta}{data.frame (t, n) containing the metadata.}
}
\description{
Plotting centroids
}
