% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_validation.R
\name{plot_model_explorer}
\alias{plot_model_explorer}
\title{Plot model explorer}
\usage{
plot_model_explorer(labels)
}
\arguments{
\item{labels}{list of labels}
}
\description{
Plot model explorer
}
