% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{pvalues_fisher_method}
\alias{pvalues_fisher_method}
\title{Fisher's method to combine pvalues}
\usage{
pvalues_fisher_method(pvalues)
}
\arguments{
\item{pvalues}{pvalues}
}
\description{
Combines all p-value per rows.
}
