% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{rescale_values}
\alias{rescale_values}
\title{Rescales centroids and gene expresion values}
\usage{
rescale_values(y, meta = NULL, group = NULL)
}
\arguments{
\item{y}{The matrix to rescale. Each row should correspond to a gene or a
centroid and columns to samples.}

\item{meta, }{optional
Metadata data.frame.}

\item{group, }{optional, default: NULL
A column name of the metadata data.frame. The corresponding column
should be factors. If provided, the values of y will be rescaled such
that, for each row, all values associated to group A, … of column
"group" of the metadata is between 0 and 1. For example, if column
"group" corresponds to a genotype, all the values of a gene for a
specific genotype will be rescaled between 0 and 1.}
}
\value{
rescaled y, such that for each row, the values are comprised
     between 0 and 1. Note that if "group" is provided, the values
     associated to the columns of unique values of "group" will be rescaled
     separately.
}
\description{
Rescales centroids and gene expresion values
}
