% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoemaker2015.R
\docType{data}
\name{shoemaker2015}
\alias{shoemaker2015}
\title{Time course transcriptomic from mouse lung tissues infected with influenza}
\format{An object of class \code{list} of length 2.}
\usage{
data(shoemaker2015)
}
\description{
Time-course microarray data from "An Ultrasensitive Mechanism Regulates
Influenza Virus-Induced Inflammation"
}
\examples{
data(shoemaker2015)
data = attr(shoemaker2015, "data")
meta = attr(shoemaker2015, "meta")
}
\references{
Shoemaker et al. (2015) PLOS Pathog.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26046528}{PubMed})
}
\keyword{datasets}
