% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{splines_kmeans}
\alias{splines_kmeans}
\title{Performs splines clustering using K-means}
\usage{
splines_kmeans(
  data,
  moanin_model,
  n_clusters = 10,
  init = "kmeans++",
  n_init = 10,
  max_iter = 300,
  random_seed = NULL,
  fit_splines = TRUE,
  rescale = TRUE
)
}
\arguments{
\item{data}{matrix containg the data. Data needs to be in log scale.}

\item{moanin_model}{moanin_model}

\item{n_clusters}{int optional, default: 10}

\item{init}{["kmeans++", "random", "optimal_init"]}

\item{n_init}{int, optional, default: 10
Number of initialization to perform.}

\item{max_iter}{int, optional, default: 300
Maximum number of iteration to perform}

\item{random_seed}{int, optional, default: NULL}

\item{fit_splines}{boolean, optional, default: TRUE
Whether to fit splines or not.}

\item{rescale}{boolean, optional, default: TRUE
Whether to rescale the data or not.}
}
\description{
Performs splines clustering using K-means
}
