% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{splines_kmeans_score_and_label}
\alias{splines_kmeans_score_and_label}
\title{Assign score and labels from raw data}
\usage{
splines_kmeans_score_and_label(
  data,
  kmeans_clusters,
  percentage_genes_to_label = 0.5,
  max_score = NULL,
  rescale_separately_on = NULL
)
}
\arguments{
\item{data}{The data matrix, where genes (features) are on the rows and
samples on the columns.}

\item{kmeans_clusters}{list of list
List returned by moanin::splines_kmeans}

\item{percentage_genes_to_label}{float, optional, default: 0.5
Percentage of genes to label. If max_score is provided, will label
genes that are either in the top `percentage_genes_to_label` or with a
score below `max_score`.}

\item{max_score}{optional, default: Null
When provided, will only label genes below that score. If NULL, ignore
this option.}

\item{rescale_separately_on, }{string, optional, default: NULL
When provided, will rescale separately different groups of data.}
}
\value{
A list consisting of
\itemize{
\item{\code{labels}}{the label or cluster assigned to each gene based on the
cluster with the best (i.e. lowest) score, with no label given to genes that
do not have a score lower than a specified quantity}
\item{\code{scores}}{the matrix of size n_cluster x n_genes, containing for 
each gene and each cluster, the goodness of fit score}
\item{\code{score_cutoff}}{The required cutoff for a gene receiving an
assignment}
 }
}
\description{
Assign score and labels from raw data
}
