% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{create_meta_prediction}
\alias{create_meta_prediction}
\title{Create prediction meta data from splines model}
\usage{
create_meta_prediction(moanin_model, num_timepoints = 100)
}
\arguments{
\item{moanin_model}{object of class \code{moanin_model}, an object containing
all related information to the splines model used. See
\code{\link{create_moanin_model}} for more details.}

\item{num_timepoints}{integer, optional, default: 100
Number of timepoints to use for the prediction metadata}
}
\description{
Create prediction meta data from splines model
}
\keyword{internal}
