% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moanin_model.R
\name{create_moanin_model}
\alias{create_moanin_model}
\alias{print.moanin_model}
\title{Create a moanin model}
\usage{
create_moanin_model(
  meta,
  formula = NULL,
  basis = NULL,
  group_variable = "Group",
  time_variable = "Timepoint",
  degrees_of_freedom = NULL
)

\method{print}{moanin_model}(x, ...)
}
\arguments{
\item{meta}{\code{data.frame} containing the metadata in columns, and rows
corresponding to different samples.}

\item{formula}{formula object, optional, default: NUlL. Used to construct
splines from the data in \code{meta}. See details.}

\item{basis}{matrix, optional, default: NULL. A basis matrix, where each row
corresponds to the evaluation of a sample on the basis function (thus one
column for each basis function).}

\item{group_variable}{A character value giving the column that corresponds to
the grouping variable to test for DE. By default "Group"}

\item{time_variable}{A character value giving the column that corresponds to
the time variable. By default "Timepoint".}

\item{degrees_of_freedom}{int, optional. Number of degrees of freedom to use
if neither the basis nor the formula is provided. If not provided by the
user, internally will be set to 4}

\item{x}{a \code{moanin_model} object}

\item{...}{arguments passed to generic \code{print}}
}
\value{
An object of class \code{moanin_model}; a list with the following elements:
\itemize{
\item{\code{basis}}{A basis matrix, where each row
 corresponds to the evaluation of a sample on the basis function (thus one
 column for each basis function). }
\item{\code{meta}}{The data frame provided by user to \code{meta}}
\item{\code{degrees_of_freedom}}{The degrees of freedom provided by the user 
(or set by the function internally)}
\item{\code{formula}}{The formula used to construct the basis matrix 
(unless basis matrix is provided by the user).}
}
}
\description{
This model is used to conserve information relating to the model, such as
the formula, the basis, the meta data.
}
\details{
If neither \code{formula} nor \code{basis} is given, then by default,
 the function will create a basis matrix based on the formula:
 \preformatted{formula = ~Group:ns(Timepoint, df=4) + Group + 0}

Note that the meta data will have levels dropped (via \code{droplevels}).
}
\examples{
# Load some data
data(exampleData)

# Use the default options
moanin = create_moanin_model(testMeta)
print(moanin)

# Change the number of degrees of freedom
moanin = create_moanin_model(testMeta, degrees_of_freedom=6)
print(moanin)
}
