% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moanin_model.R
\name{create_moanin_model}
\alias{create_moanin_model}
\title{Create a moanin model}
\usage{
create_moanin_model(meta, formula = NULL, basis = NULL,
  degrees_of_freedom = 4)
}
\arguments{
\item{meta}{data.frame containing the metadata. Metadata needs to contain
column "Group" and "Timepoint"}

\item{formula}{formula object, optional, default: NUlL
Formula of the splines, e.g:
    formula = ~Group:splines::ns(Timepoint, df=4) + Group + 0}

\item{basis}{matrix, optional, default: NULL
A basis object.}

\item{degrees_of_freedom}{int, optional, default: 4
Number of degrees of freedom to use if neither the basis nor the
formula is provided}
}
\description{
This model is used to conserve information relating to the model, such as
the formula, the basis, the meta data.
}
