% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moanin_model.R
\name{create_splines_model}
\alias{create_splines_model}
\title{Depricated function for creating a moanin_model object}
\usage{
create_splines_model(
  meta,
  formula = NULL,
  basis = NULL,
  degrees_of_freedom = 4
)
}
\arguments{
\item{inheritParams}{create_moanin_model}
}
\description{
This function is depricated. Users should use \code{\link{create_moanin_model}}
}
\details{
This function is depricated. Users should use \code{\link{create_moanin_model}}
}
\keyword{internal}
