% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_timecourse.R
\name{DE_timecourse}
\alias{DE_timecourse}
\title{Run edge with contrasts.}
\usage{
DE_timecourse(
  data,
  moanin_model,
  contrasts,
  center = FALSE,
  use_voom_weights = TRUE
)
}
\arguments{
\item{data}{The data matrix.}

\item{moanin_model}{moanin_model
Object containing all related information to the splines model used.}

\item{contrasts}{Contrasts, either provided as a vector of strings, or a
matrix of contrasts coefficients obtained using \code{makeContrasts}
from \code{limma}.}

\item{center}{boolean, whether to center the data matrix}

\item{use_voom_weights}{boolean, optional, default: TRUE
Whether to use voom weights.}
}
\description{
Run edge with contrasts.
}
