% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_timecourse.R
\name{DE_timecourse}
\alias{DE_timecourse}
\title{Run spline models with contrasts.}
\usage{
DE_timecourse(
  data,
  moanin_model,
  contrasts,
  center = FALSE,
  use_voom_weights = TRUE
)
}
\arguments{
\item{data}{The data matrix, where genes (features) are on the rows and
samples on the columns.}

\item{moanin_model}{object of class \code{moanin_model}, an object containing
all related information for time course data and the splines model that
will be used (if applicable). See \code{\link{create_moanin_model}} for
more details.}

\item{contrasts}{Contrasts, either provided as a vector of strings, or a
matrix of contrasts coefficients obtained using
\code{\link[limma]{makeContrasts}} from the package \code{limma}. If given
as a character string, will be passed to \code{\link[limma]{makeContrasts}}
to be converted into such a matrix.}

\item{center}{boolean, whether to center the data matrix}

\item{use_voom_weights}{boolean, optional, default: TRUE. 
Whether to use voom weights.}
}
\value{
A \code{data.frame} with two columns for each of the contrasts given
  in \code{contrasts}, corresponding to the raw p-value of the contrast for
  that gene (\code{_pval}) and the adjusted p-value (\code{_qval}). The
  adjusted p-values are FDR-adjusted based on the Benjamini-Hochberg method,
  as implemented in \code{\link[stats]{p.adjust}}. The adjustment is done
  across all p-values for all contrasts calculated.
}
\description{
Run spline models with contrasts.
}
\details{
The implementation of the spline fit and the calculation of p-values
  was based on code from \code{\link[edge]{edge}}, and expanded to
  enable handling of comparisons of groups via contrasts.#' @seealso \code{\link[limma]{makeContrasts}}, \code{\link{create_moanin_model}}, \code{\link{DE_timepoints}}, \code{\link[edge]{edge}}
}
\examples{
data(exampleData)
moanin = create_moanin_model(testMeta)
deTimecourse=DE_timecourse(data=testData, moanin_model=moanin, contrasts="K-C", use_voom_weights=FALSE)
head(deTimecourse)
}
