% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_timepoints.R
\name{DE_timepoints}
\alias{DE_timepoints}
\title{Fit weekly differential expression analysis}
\usage{
DE_timepoints(data, moanin_model, contrasts, use_voom_weights = TRUE)
}
\arguments{
\item{data}{The data matrix, where genes (features) are on the rows and
samples on the columns.}

\item{moanin_model}{object of class \code{moanin_model}, an object containing
all related information to the splines model used. See
\code{\link{create_moanin_model}} for more details.}

\item{contrasts}{Contrasts, either provided as a vector of strings, or a
matrix of contrasts coefficients obtained using \code{makeContrasts}
from \code{limma}.}

\item{use_voom_weights}{boolean, optional, default: TRUE. 
Whether to use voom weights.}
}
\description{
Fit weekly differential expression analysis
}
