% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_timepoints.R
\name{DE_timepoints}
\alias{DE_timepoints}
\alias{create_timepoints_contrasts}
\alias{DE_timepoints,Moanin-method}
\alias{create_timepoints_contrasts,Moanin-method}
\title{Fit weekly differential expression analysis}
\usage{
\S4method{DE_timepoints}{Moanin}(object, contrasts, add_factors = NULL, use_voom_weights = TRUE)

\S4method{create_timepoints_contrasts}{Moanin}(
  object,
  group1,
  group2,
  type = c("per_timepoint", "diff_timepoint"),
  timepoints = sort(unique(time_variable(object))),
  timepoints_before = head(sort(timepoints), -1),
  timepoints_after = tail(sort(timepoints), -1)
)
}
\arguments{
\item{object}{An object of class \code{\link{Moanin}}, an object containing
all related information for time course data and the splines model that
will be used (if applicable). See \code{\link{create_moanin_model}} for
more details.}

\item{contrasts}{Contrasts, either provided as a vector of strings, or a
matrix of contrasts coefficients obtained using
\code{\link[limma]{makeContrasts}} from the package \code{limma}. If given
as a character string, will be passed to \code{\link[limma]{makeContrasts}}
to be converted into such a matrix.}

\item{add_factors}{A character vector of additional variables to add to the 
design. See details.}

\item{use_voom_weights}{boolean, optional, default: TRUE. 
Whether to use voom weights. See details.}

\item{group1}{First group to consider in making contrasts, character value
that must match a value contained in \code{moanin_model$meta}.}

\item{group2}{Second group to consider in making contrasts, character value
that must match a value contained in \code{moanin_model$meta}.}

\item{timepoints_before}{for create_diff_contrasts, see details the set of timepoints to compare, where contrast will be timepoints_after - timepoints_before}

\item{timepoints_after}{for create_diff_contrasts, the set of timepoints to compare, see details where contrast will be timepoints_after - timepoints_before}
}
\value{
A \code{data.frame} with three columns for each of the contrasts
  given in \code{contrasts}, corresponding to the raw p-value of the contrast
  for that gene (\code{_pval}), the adjusted p-value (\code{_qval}), and the
  estimate of log-fold-change (\code{_lfc}). The adjusted p-values are
  FDR-adjusted based on the Benjamini-Hochberg method, as implemented in
  \code{\link[stats]{p.adjust}}. The adjustment is done across all p-values
  for all contrasts calculated.

\code{create_timepoints_contrasts}: a character vector with each
  element of the vector corresponding to a contrast to be compared.
}
\description{
Fit weekly differential expression analysis

Creates pairwise contrasts for all timepoints
}
\details{
By default the formula fitted for each gene is
\preformatted{   
   ~ Group*Timepoint +0
}
If the user gives values to \code{add_factors}, then the vector of character 
values given in \code{add_factors} will be \emph{added} to the default formula. 
So that \code{add_factors="Replicate"} will change the formula to
\preformatted{   
   ~ Group*Timepoint +0 + Replicate
}
This allows for a small amount of additional complexity to control 
for other variables. Users should work directly with limma for 
more complex models.

If \code{use_voom_weights=TRUE}, the data is given directly to limma
  via \code{assay(object)}. The specific series of
  calls is:
\preformatted{   
   y <- edgeR::DGEList(counts=assay(object))
   y <- edgeR::calcNormFactors(y, method="upperquartile")
   v <- limma::voom(y, design, plot=FALSE)
   v <- limma::lmFit(v) 
   }

If the user set \code{log_transform=TRUE} in the creation of the
  \code{Moanin} object, this will not have an impact in the analysis if
  \code{use_voom_weights=TRUE}. Only if \code{use_voom_weights=FALSE} will
  this matter, in which case the log of the input data will be given to a
  regular call to \code{limma}:
\preformatted{
   y<-get_log_data(object)
   v <- limma::lmFit(y, design)
}

\code{create_timepoints_contrasts} creates the needed contrasts for
  comparing two groups for every timepoint in the format needed for
  \code{DE_timepoints} (i.e. \code{\link[limma]{makeContrasts}}, to which the
  contrasts are ultimately passed). The time points are determined by the
  meta data in the \code{moanin_object} provided by the user.

\code{type="diff_timepoint"} will create contrasts that compare the difference between two timepoints between two levels of the \code{group_variable} of the \code{Moanin} object. These are contrasts in the form (TP i - TP (i-1))[Group1] - (TP i - TP (i-1))[Group2].
}
\examples{
data(exampleData)
moanin <- create_moanin_model(data=testData, meta=testMeta)
contrasts <- create_timepoints_contrasts(moanin,"C", "K")
head(contrasts)
deTimepoints=DE_timepoints(moanin, 
    contrasts=contrasts, use_voom_weights=FALSE)
head(deTimepoints)
contrastsDiff <- create_diff_contrasts(moanin,"C", "K")
deDiffTimepoints=DE_timepoints(moanin, 
    contrasts=contrasts,
    use_voom_weights=FALSE)
# provide the sets of timepoints to compare:
contrastsDiff<-create_timepoints_contrasts(moanin,"C", "K",timepoints_before=c(72,120),timepoints_after=c(168,168),type="diff_timepoint")
# Control for replicate variable:
deTimepoints=DE_timepoints(moanin, 
    contrasts=contrasts, add_factors="Replicate",
    use_voom_weights=FALSE)
head(deTimepoints)
}
\seealso{
\code{\link[limma]{makeContrasts}}
}
