% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis_utils.R
\name{estimate_log_fold_change}
\alias{estimate_log_fold_change}
\title{Estimates log fold change}
\usage{
estimate_log_fold_change(data, moanin_model, contrasts,
  method = "timecourse")
}
\arguments{
\item{data}{The data in a matrix}

\item{moanin_model}{}

\item{contrasts}{The contrasts to consider}

\item{method}{["sum", "max", "min", "timely", "timecourse", "epicon"]}
}
\description{
Estimates log fold change
}
