% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis_utils.R
\name{estimate_log_fold_change}
\alias{estimate_log_fold_change}
\title{Estimates log fold change}
\usage{
estimate_log_fold_change(
  data,
  moanin_model,
  contrasts,
  method = c("timecourse", "sum", "max", "timely", "epicon", "abs_sum",
    "abs_squared_sum", "min")
)
}
\arguments{
\item{data}{The data matrix, where genes (features) are on the rows and
samples on the columns.}

\item{moanin_model}{object of class \code{moanin_model}, an object containing
all related information to the splines model used. See
\code{\link{create_moanin_model}} for more details.}

\item{contrasts}{The contrasts to consider}

\item{method}{method for calculating the log-fold change. See details.}
}
\description{
Estimates log fold change
}
\details{
The following methods exist for calculating the log-fold change
  between conditions over time (default is "timecourse"):
\itemize{
\item{\code{timecourse}}{}
\item{\code{sum}}{}
\item{\code{max}}{}
\item{\code{timely}}{}
\item{\code{epicon}}{}
\item{\code{abs_sum}}{}
\item{\code{abs_squared_sum}}{}
\item{\code{min}}{}
}
}
