% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_terms_analysis.R
\name{find_enriched_go_terms}
\alias{find_enriched_go_terms}
\title{Find enriched GO terms}
\usage{
find_enriched_go_terms(assignments, gene_id_to_go, ontology = "BP",
  weighted = FALSE, node_size = 10)
}
\arguments{
\item{assignments}{boolean named vector}

\item{gene_id_to_go}{Gene ID to GO object from topGO}

\item{ontology}{string, optional, default: BP
specficies which ontology to use. Can be 'BP', 'CC', or 'NF'}

\item{weighted, }{boolean, optional, default: FALSE
Whether to use the weighted algorithm or not.}

\item{node_size}{integer, optional, default: 10
Consider only GO terms with node_size number of genes.}

\item{ontology:}{BP, CC, NF}
}
\description{
Find enriched GO terms
}
