% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{fit_predict_splines}
\alias{fit_predict_splines}
\title{Get fitted values for splines for each gene}
\usage{
fit_predict_splines(data, moanin_model, meta_prediction = NULL)
}
\arguments{
\item{data}{The data matrix, where genes (features) are on the rows and
samples on the columns.}

\item{moanin_model}{object of class \code{moanin_model}, an object containing
all related information for time course data and the splines model that
will be used (if applicable). See \code{\link{create_moanin_model}} for
more details.}

\item{meta_prediction}{}
}
\value{
a matrix of the fitted y values, with dimensions the same as \code{data}
}
\description{
Get fitted values for splines for each gene
}
\keyword{internal}
