% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{fit_splines}
\alias{fit_splines}
\title{Fit splines}
\usage{
fit_splines(data, moanin_model, weights = NULL)
}
\arguments{
\item{data}{The data matrix, where genes (features) are on the rows and
samples on the columns.}

\item{moanin_model}{object of class \code{moanin_model}, an object containing
all related information to the splines model used. See
\code{\link{create_moanin_model}} for more details.}

\item{weights}{weigts}
}
\value{
beta coefficients
}
\description{
Fit splines
}
