% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{fit_splines}
\alias{fit_splines}
\title{Fit splines to each gene of data matrix}
\usage{
fit_splines(moanin_model, data = NULL, weights = NULL)
}
\arguments{
\item{moanin_model}{object of class Moanin}

\item{data}{a matrix of data to fix splines to. If NULL, uses
\code{assay(moanin_model)}}

\item{weights}{A matrix of weights, of the same dimension as \code{data}.}
}
\value{
matrix of the coefficients for each basis function, each row of the
  matrix containing the coefficients for the corresponding gene in
  \code{data}.
}
\description{
Fit splines to each gene of data matrix
}
\keyword{internal}
