% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{fit_splines}
\alias{fit_splines}
\title{Fit splines to each gene of data matrix, used by DE_timecourse}
\usage{
fit_splines(design_matrix, data, weights = NULL)
}
\arguments{
\item{design_matrix}{design matrix (containing evaluated basis matrix, but potentially other variables) to fit}

\item{data}{a matrix of data to fix splines to}

\item{weights}{A matrix of weights, of the same dimension as \code{data}.}
}
\value{
matrix of the coefficients for each basis function, each row of the
  matrix containing the coefficients for the corresponding gene in
  \code{data}.
}
\description{
Fit splines to each gene of data matrix, used by DE_timecourse
}
\details{
Needed to allow for fitting weights, in which case for loop over every row/gene. Otherwise, just a call to lm.fit.
}
\keyword{internal}
