% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_validation.R
\name{get_auc_similarity_scores}
\alias{get_auc_similarity_scores}
\title{Plot AUC consensus matrix}
\usage{
get_auc_similarity_scores(labels, method = c("consensus", "nmi"))
}
\arguments{
\item{labels}{a list. Each element of
the list is a matrix that gives the results of a clustering routine in each
column (see \code{\link{consensus_matrix}}). Usually each column would be
the result of running the clustering on a subsample or bootstrap resample
of the data.}

\item{method}{method for calculation of similarity, one of "consensus" or "nmi". See details.}
}
\description{
Plot AUC consensus matrix
}
