% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_validation.R
\name{plot_model_explorer}
\alias{plot_model_explorer}
\title{Plot model explorer}
\usage{
plot_model_explorer(labels)
}
\arguments{
\item{labels}{a list. Each element of
the list is a matrix that gives the results of a clustering routine in each
column (see \code{\link{consensus_matrix}}). Usually each column would be
the result of running the clustering on a subsample or bootstrap resample
of the data.}
}
\description{
Plot model explorer
}
