% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_splines_data}
\alias{plot_splines_data}
\title{Plotting splines}
\usage{
plot_splines_data(data, splines_model, colors = NULL, smooth = FALSE,
  legend = TRUE, legendArgs = NULL, subset_conditions = NULL,
  subset_data = NULL, simpleY = TRUE, mar = c(2.5, 2.5, 3, 1),
  mfrow = NULL, addToPlot = NULL, ...)
}
\arguments{
\item{data}{matrix (k, t) containing the data (such as centroids or data)}

\item{splines_model}{splines_model}

\item{colors}{vector, optional, default NULL
vector of colors}

\item{smooth}{boolean, optional, default: FALSE
Whether to smooth the centroids or not.}

\item{legend}{boolean whether to include a legend (default:TRUE)}

\item{legendArgs}{list of arguments to be passed to legend command
(if \code{legend=TRUE})}

\item{subset_conditions}{list
if provided, only plots the subset of conditions provided. Else, plots
all conditions}

\item{subset_data}{list
if provided, only plots  the subset of data (ie, the rows) provided.}

\item{simpleY}{boolean, if true, will minimize the annotation of the
y axis to  only label the axis in the exterior plots (the x-axis
is always assumed to be the same across all plots and will always
 be simplified)}

\item{mar}{vector of margins to set the space around each plot
(see \code{\link{par}})}

\item{mfrow}{a vector of integers of length 2 defining the grid of
plots to be created (see \code{\link{par}}). If missing, the
function will set a value.}

\item{...}{arguments to be passed to the individual plot commands
(Will be sent to all plot commands)}

\item{meta}{data.frame (t, n) containing the metadata.}
}
\description{
Plotting splines
}
