% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_splines_data}
\alias{plot_splines_data}
\title{Plotting splines}
\usage{
plot_splines_data(
  data,
  moanin_model,
  colors = NULL,
  smooth = FALSE,
  legend = TRUE,
  legendArgs = NULL,
  subset_conditions = NULL,
  subset_data = NULL,
  simpleY = TRUE,
  mar = c(2.5, 2.5, 3, 1),
  mfrow = NULL,
  addToPlot = NULL,
  ...
)
}
\arguments{
\item{data}{matrix containing the data to be plotted (such as centroids or
data), where each row of the data provided will be plotted as a separate
plot.}

\item{moanin_model}{object of class \code{moanin_model}, an object containing
all related information for time course data and the splines model that
will be used (if applicable). See \code{\link{create_moanin_model}} for
more details.}

\item{colors}{vector, optional, default NULL. Vector of colors}

\item{smooth}{boolean, optional, default: FALSE. Whether to smooth the
centroids or not.}

\item{legend}{boolean whether to include a legend (default:TRUE)}

\item{legendArgs}{list of arguments to be passed to legend command (if
\code{legend=TRUE})}

\item{subset_conditions}{list if provided, only plots the subset of conditions
provided. Else, plots all conditions}

\item{subset_data}{list if provided, only plots  the subset of data (ie, the
rows) provided.}

\item{simpleY}{boolean, if true, will minimize the annotation of the y axis to
only label the axis in the exterior plots (the x-axis is always assumed to
be the same across all plots and will always be simplified)}

\item{mar}{vector of margins to set the space around each plot (see
\code{\link{par}})}

\item{mfrow}{a vector of integers of length 2 defining the grid of plots to be
created (see \code{\link{par}}). If missing, the function will set a value.}

\item{addToPlot}{A function that will be called after the plotting, allowing
the user to add more to the plot.}

\item{...}{arguments to be passed to the individual plot commands
(Will be sent to all plot commands)}
}
\description{
Plotting splines
}
\examples{
# First, load some data and create a moanin model
data(exampleData)
moanin_model = create_moanin_model(testMeta, degrees_of_freedom=6)

# The moanin model contains all the information for plotting purposes. The
# plot_splines_function will automatically fit the splines from the
# information contained in the moanin model
plot_splines_data(
testData, moanin_model,
subset_data=c("AK050122", "AK043921"),
mfrow=c(2, 2))

# The splines can also be smoothed
plot_splines_data(testData, moanin_model,
	     subset_data=c("AK050122", "AK043921"),
	     smooth=TRUE, mfrow=c(2, 2))
}
