% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{splines_kmeans}
\alias{splines_kmeans}
\title{Performs splines clustering using K-means}
\usage{
splines_kmeans(
  data,
  moanin_model,
  n_clusters = 10,
  init = "kmeans++",
  n_init = 10,
  max_iter = 300,
  random_seed = NULL,
  fit_splines = TRUE,
  rescale = TRUE
)
}
\arguments{
\item{data}{The data matrix, where genes (features) are on the rows and
samples on the columns.}

\item{moanin_model}{object of class \code{moanin_model}, an object containing
all related information for time course data and the splines model that
will be used (if applicable). See \code{\link{create_moanin_model}} for
more details.}

\item{n_clusters}{int optional, default: 10}

\item{init}{["kmeans++", "random", "optimal_init"]}

\item{n_init}{int, optional, default: 10
Number of initialization to perform.}

\item{max_iter}{int, optional, default: 300
Maximum number of iteration to perform}

\item{random_seed}{int, optional, default: NULL}

\item{fit_splines}{boolean, optional, default: TRUE
Whether to fit splines or not.}

\item{rescale}{boolean, optional, default: TRUE
 Whether to rescale the data or not.
 @return A list in the format returned by \code{\link[ClusterR]{KMeans_rcpp}},
   with the following elements added or changed:
\itemize{
\item{\code{centroids}}{The centroids are rescaled so that they range from 0-1}
\item{\code{moanin_model}}{The given moanin_model}
\item{\code{fit_splines}}{The value of \code{fit_splines} given to the function }
\item{\code{rescale}}{The value of \code{rescale} given to the function }
}}
}
\description{
Performs splines clustering using K-means
}
