% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{splines_kmeans_score_and_label}
\alias{splines_kmeans_score_and_label}
\title{Assign score and labels from raw data}
\usage{
splines_kmeans_score_and_label(data, kmeans_clusters,
  percentage_genes_to_label = 0.5)
}
\arguments{
\item{data}{array or dataframe.
Data to assign score and labels to}

\item{kmeans_clusters}{list of list
List returned by moanin::splines_kmeans}

\item{percentage_genes_to_label}{float, optional, default: 0.5
Percentage of genes to label.}
}
\description{
Assign score and labels from raw data
}
